requirements.txt
requests~=2.32.3
pillow~=10.4.0
transformers~=4.43.4
torch~=2.4.0+cu124